Tutorial Step 11
Let's replace the hard-coded data with some dynamic data from the
server. We will remove the data prop and replace it with a URL to fetch:

// tutorial11.js
ReactDOM.render(
<CommentBox url="/api/comments" />,
document.getElementById('content')
);

Code
This component is different from the prior components because it will
have to re-render itself. The component won't have any data until the
request from the server comes back, at which point the component may
need to render some new comments.

Note: the code will not be working at this step.